# Optimize `fbuffer_append_str_repeat`